requirements.txt
# Core
python-dotenv
pandas
numpy
pyyaml
openai
google-generativeai

# Optional AI orchestration
langchain
langchain-openai
langchain-google-genai

# Visualization / logs / testing
matplotlib
pytest
tqdm

# Optional structured logging / tracing
langfuse